1. most recently entered samples are now on top 2. don't call cleanup when there is no selected project
git-svn-id: svn://svn.code.sf.net/p/biominerqt/code-0/trunk@206 86439291-5e3c-406c-9808-073398cf6135

+++ src/main/webapp/app/submit/UploadController.js
@@ -849,13 +849,16 @@ angular.module("upload").controller("UploadController", ['$scope','$upload','$ht
 		 *  Delete all failed/incomplete on back end
 		 */
 		$scope.cleanUploadedFiles = function() {
-			return $http({
-				url: "submit/cleanUploadedFiles",
-				method: "DELETE",
-				params: {idProject: $scope.projectId}
-			}).success(function(data) {
-				updateFileList(data);
-			});
+			if ($scope.projectId != -1) {
+				return $http({
+					url: "submit/cleanUploadedFiles",
+					method: "DELETE",
+					params: {idProject: $scope.projectId}
+				}).success(function(data) {
+					updateFileList(data);
+				});
+			}
+			
 		}
 		
 		/****************
